FiniteElementContainers

Documentation for FiniteElementContainers.

FiniteElementContainers.FileMesh

FiniteElementContainers.StaticAssembler

FiniteElementContainers.assemble!

FiniteElementContainers.assemble!

FiniteElementContainers.assemble!

FiniteElementContainers.bc_ids

FiniteElementContainers.file_name

FiniteElementContainers.update_unknown_ids!

SparseArrays.sparse

FiniteElementContainers.FileMesh — Type

file_name::String

mesh_obj::Any

Mesh type that has a handle to an open mesh file object. This type's methods are "overridden" in extensions.

See FiniteElementContainersExodusExt for an example.

FiniteElementContainers.StaticAssembler — Method

Default initialization Assumes no dirichlet bcs

TODO add typing to constructor

FiniteElementContainers.assemble! — Method

assemble!(
    assembler::StaticAssembler,
    dof::DofManager,
    fspace::FunctionSpace,
    X,
    U,
    block_id,
    residual_func,
    tangent_func
)

Simple method for assembling in serial

assemble!(
    assembler,
    dof,
    fspace,
    X,
    U,
    block_id,
    residual_func,
    tangent_func
)

defined at /home/runner/work/FiniteElementContainers.jl/FiniteElementContainers.jl/src/Assemblers.jl:118.

FiniteElementContainers.assemble! — Method

assemble!(
    assembler::StaticAssembler,
    K_el::AbstractMatrix,
    block_id::Int64,
    el_id::Int64
) -> AbstractMatrix

assembly for stiffness matrix

assemble!(assembler, K_el, block_id, el_id)

defined at /home/runner/work/FiniteElementContainers.jl/FiniteElementContainers.jl/src/Assemblers.jl:100.

FiniteElementContainers.assemble! — Method

assemble!(
    assembler::StaticAssembler,
    R_el::AbstractVector,
    conn
)

assembly method for just a residual vector

TODO need to add an Atomix lock here TODO add block_id to fspace or something like that

assemble!(assembler, R_el, conn)

defined at /home/runner/work/FiniteElementContainers.jl/FiniteElementContainers.jl/src/Assemblers.jl:86.

FiniteElementContainers.bc_ids — Method

bc_ids(dof_manager::DofManager, nsets, dofs) -> Any

eventually use this method

bc_ids(dof_manager, nsets, dofs)

defined at /home/runner/work/FiniteElementContainers.jl/FiniteElementContainers.jl/src/DofManagers.jl:115.

FiniteElementContainers.file_name — Method

file_name(mesh::FiniteElementContainers.AbstractMesh) -> Any

Returns file name for an mesh type

file_name(mesh)

defined at /home/runner/work/FiniteElementContainers.jl/FiniteElementContainers.jl/src/Meshes.jl:18.

FiniteElementContainers.update_unknown_ids! — Method

update_unknown_ids!(
    assembler::StaticAssembler,
    fspaces,
    nodes_in
)

method that assumes first dof TODO move sorting of nodes up stream TODO remove other scratch unknowns and unknown_indices arrays

update_unknown_ids!(assembler, fspaces, nodes_in)

defined at /home/runner/work/FiniteElementContainers.jl/FiniteElementContainers.jl/src/Assemblers.jl:174.

SparseArrays.sparse — Method

sparse(
    assembler::StaticAssembler
) -> SparseArrays.SparseMatrixCSC

sparse(assembler)

defined at /home/runner/work/FiniteElementContainers.jl/FiniteElementContainers.jl/src/Assemblers.jl:74.