Connectivities

There are a few types that are aliases of ElementFields that are related to connectivity arrays in finite elements.

FiniteElementContainers.Connectivity — Type

abstract type ElementField{T, N, NN, NE, Vals} <: FiniteElementContainers.AbstractField{T, N, NN, Vals}

FiniteElementContainers.SimpleConnectivity — Type

struct SimpleElementField{T, N, NN, NE, Vals} <: ElementField{T, N, NN, NE, Vals}

FiniteElementContainers.VectorizedConnectivity — Type

struct VectorizedElementField{T, N, NN, NE, Vals} <: ElementField{T, N, NN, NE, Vals}

The main method for these types is connectivity

FiniteElementContainers.connectivity — Function

connectivity(
    conn::FiniteElementContainers.SimpleElementField
) -> Any

connectivity(
    conn::FiniteElementContainers.SimpleElementField,
    e::Int64
) -> Any

connectivity(
    conn::FiniteElementContainers.VectorizedElementField{T, N, NN, NE, Vals} where {T, N, NN, NE, Vals}
) -> Any

connectivity(
    conn::FiniteElementContainers.VectorizedElementField{T, 1, NN, NE, Vals},
    e::Int64
) -> Any

connectivity(
    conn::FiniteElementContainers.VectorizedElementField{T, 2, NN, NE, Vals},
    e::Int64
) -> SubArray{T, 1} where T

connectivity(fspace::FunctionSpace) -> Any

connectivity(fspace::FunctionSpace, e::Int64) -> Any