Fields

Fields serve as loose wrappers around AbstractArray subtypes such that the size of array slices are known at compile time. Although this introduces a type-instability, the idea is to do this at the top most level (mainly at setup time of a FEM simulation). By introducing this type instability, we can gain information about the field type that is used in methods downstream to construct StaticArrays of views of field types.

All fields are subtypes of the abstract type AbstractField

julia> using FiniteElementContainers
julia> FiniteElementContainers.AbstractFieldFiniteElementContainers.AbstractField

Example - H1Field a.k.a. NodalField

We can set up a H1Field in one of two ways. The simplest constructor form can be used as follows

julia> using FiniteElementContainers
julia> field = H1Field(rand(2, 10))2×10 H1Field{Float64, Vector{Float64}, 2}:
 0.546121  0.599658   0.0869196  0.0182491  …  0.553101  0.935078  0.341752
 0.860652  0.0714353  0.605594   0.875427      0.839182  0.4117    0.158346

This is stored in a vectorized way as can be seen above

julia> field.data20-element Vector{Float64}:
 0.5461214210539747
 0.8606524661567391
 0.5996583374920754
 0.07143533284193393
 0.08691964391758622
 0.6055943390800693
 0.018249098831686128
 0.8754266549308382
 0.09936019372730753
 0.543738698014828
 0.08359560501386831
 0.15020286589679688
 0.6730776956184814
 0.6372127333795637
 0.5531012257795784
 0.839182299521225
 0.9350784989253339
 0.4117003759392043
 0.34175186088757825
 0.1583463344592536

Fields can be indexed like regular arrays, e.g.

julia> field[1, 1]0.5461214210539747

julia> field[1, :]10-element Vector{Float64}:
 0.5461214210539747
 0.5996583374920754
 0.08691964391758622
 0.018249098831686128
 0.09936019372730753
 0.08359560501386831
 0.6730776956184814
 0.5531012257795784
 0.9350784989253339
 0.34175186088757825

etc.

Fields

The base type for fields is the AbstractField abstract type.

Base.IndexStyle — Method

IndexStyle(
    _::Type{<:FiniteElementContainers.AbstractField}
) -> IndexLinear

FiniteElementContainers.AbstractField — Type

abstract type AbstractField{T, N, D<:AbstractArray{T, 1}} <: AbstractArray{T, N}

Thin wrapper that subtypes AbstractArray and serves as the base Field type

FiniteElementContainers.Connectivity — Type

struct Connectivity{T<:Integer, D<:AbstractArray{T<:Integer, 1}} <: FiniteElementContainers.AbstractBlockField{T<:Integer, D<:AbstractArray{T<:Integer, 1}}

FiniteElementContainers.Connectivity_v2 — Type

struct Connectivity_v2{T<:Integer, D<:AbstractArray{T<:Integer, 1}} <: FiniteElementContainers.AbstractBlockField{T<:Integer, D<:AbstractArray{T<:Integer, 1}}

FiniteElementContainers.H1Field — Type

struct H1Field{T, D, NF} <: FiniteElementContainers.AbstractContinuousField{T, D, NF}

Implementation of fields that live on nodes.

FiniteElementContainers.HcurlField — Type

struct HcurlField{T, D, NF} <: FiniteElementContainers.AbstractContinuousField{T, D, NF}

Implementation of fields that live in Hdiv spaces.

FiniteElementContainers.HdivField — Type

struct HdivField{T, D, NF} <: FiniteElementContainers.AbstractContinuousField{T, D, NF}

Implementation of fields that live in Hdiv spaces.

Base.eltype — Method

eltype(
    _::FiniteElementContainers.AbstractField{T, N, D}
) -> Any

Base.fill! — Method

fill!(
    field::FiniteElementContainers.AbstractField{T, N, D},
    v
) -> Any

Base.getindex — Method

getindex(
    field::FiniteElementContainers.AbstractField,
    n::Int64
) -> Any

Base.setindex! — Method

setindex!(
    field::FiniteElementContainers.AbstractField{T, N, D},
    v,
    n::Int64
)

Base.unique — Method

unique(field::FiniteElementContainers.AbstractField) -> Any

FiniteElementContainers.num_entities — Method

num_entities(
    field::FiniteElementContainers.AbstractContinuousField{T, D, NF}
) -> Any

FiniteElementContainers.num_fields — Method

num_fields(
    _::FiniteElementContainers.AbstractContinuousField{T, D, NF}
) -> Any

KernelAbstractions.get_backend — Method

get_backend(
    field::FiniteElementContainers.AbstractField
) -> Any