Parameters

FiniteElementContainers.AbstractParameters — Type

abstract type AbstractParameters

FiniteElementContainers.Parameters — Type

struct Parameters{IT<:Integer, RT<:Number, IV<:AbstractArray{IT<:Integer, 1}, RV<:AbstractArray{RT<:Number, 1}, RM1<:(AbstractMatrix), RM2<:(AbstractMatrix), RM3<:(AbstractMatrix), RM4<:(AbstractMatrix), ICFuncs<:(AbstractVector), DBCFuncs<:(AbstractVector), SRCFuncs<:(AbstractVector), NBCFuncs<:(AbstractVector), PBCFuncs<:(AbstractVector), RBCFuncs<:(AbstractVector), Phys, Props, Coords<:FiniteElementContainers.AbstractField, Field<:FiniteElementContainers.AbstractField} <: FiniteElementContainers.AbstractParameters

ics::InitialConditions{ICFuncs, IV, RV} where {IT<:Integer, RT<:Number, IV<:AbstractVector{IT}, RV<:AbstractVector{RT}, ICFuncs<:(AbstractVector)}

dirichlet_bcs::DirichletBCs{DBCFuncs, IV, RV} where {IT<:Integer, RT<:Number, IV<:AbstractVector{IT}, RV<:AbstractVector{RT}, DBCFuncs<:(AbstractVector)}

neumann_bcs::NeumannBCs{NBCFuncs, IT, IV, RM1} where {IT<:Integer, IV<:AbstractVector{IT}, RM1<:(AbstractMatrix), NBCFuncs<:(AbstractVector)}

periodic_bcs::PeriodicBCs{PBCFuncs, IV, RV} where {IT<:Integer, RT<:Number, IV<:AbstractVector{IT}, RV<:AbstractVector{RT}, PBCFuncs<:(AbstractVector)}

robin_bcs::RobinBCs{RBCFuncs, IT, IV, RM2, RM3} where {IT<:Integer, IV<:AbstractVector{IT}, RM2<:(AbstractMatrix), RM3<:(AbstractMatrix), RBCFuncs<:(AbstractVector)}

sources::Sources{SRCFuncs, RM4} where {RM4<:(AbstractMatrix), SRCFuncs<:(AbstractVector)}

times::TimeStepper

physics::Any

properties::Any

state_old::L2Field{RT, RV} where {RT<:Number, RV<:AbstractVector{RT}}

state_new::L2Field{RT, RV} where {RT<:Number, RV<:AbstractVector{RT}}

coords::FiniteElementContainers.AbstractField

field::FiniteElementContainers.AbstractField

field_old::FiniteElementContainers.AbstractField

hvp_scratch_field::FiniteElementContainers.AbstractField

FiniteElementContainers.coordinates — Method

coordinates(
    p::FiniteElementContainers.AbstractParameters
) -> Any

FiniteElementContainers.current_time — Method

current_time(
    p::FiniteElementContainers.AbstractParameters
) -> Any

FiniteElementContainers.dirichlet_dofs — Method

dirichlet_dofs(
    p::FiniteElementContainers.AbstractParameters
) -> Any

FiniteElementContainers.initialize! — Method

initialize!(p::FiniteElementContainers.AbstractParameters)

FiniteElementContainers.periodic_dofs — Method

periodic_dofs(
    p::FiniteElementContainers.AbstractParameters
) -> Tuple{Any, Any}

FiniteElementContainers.time_step — Method

time_step(
    p::FiniteElementContainers.AbstractParameters
) -> Any

FiniteElementContainers.update_bc_values! — Method

update_bc_values!(
    p::FiniteElementContainers.AbstractParameters,
    assembler
)

This method is used to update the stored bc values. This should be called at the beginning of any load step

This method only handles updating bc values for Dirichlet and Neumann BCs

Robin BC updates are handled in robin assembly method

FiniteElementContainers.update_dofs! — Method

update_dofs!(
    asm::FiniteElementContainers.AbstractAssembler,
    p::Parameters
)

FiniteElementContainers.update_time! — Method

update_time!(p::FiniteElementContainers.AbstractParameters)