FiniteElementContainers

Documentation for FiniteElementContainers.

FiniteElementContainers.Assembler

FiniteElementContainers.DynamicAssembler

FiniteElementContainers.DynamicAssembler

FiniteElementContainers.StaticAssembler

FiniteElementContainers.StaticAssembler

FiniteElementContainers.assemble!

FiniteElementContainers.assemble!

FiniteElementContainers.assemble!

FiniteElementContainers.assemble!

FiniteElementContainers.assemble!

FiniteElementContainers.bc_ids

FiniteElementContainers.file_name

FiniteElementContainers.float_type

FiniteElementContainers.int_type

FiniteElementContainers.update_unknown_dofs!

FiniteElementContainers.update_unknown_dofs!

FiniteElementContainers.update_unknown_dofs!

SparseArrays.sparse

SparseArrays.sparse

SparseArrays.sparse!

SparseArrays.sparse!

FiniteElementContainers.Assembler — Type

FiniteElementContainers.DynamicAssembler — Type

Is::AbstractVector

Js::AbstractVector

unknown_dofs::AbstractVector

block_sizes::AbstractVector

block_offsets::AbstractVector

residuals::AbstractArray

stiffnesses::AbstractVector

masses::AbstractVector

klasttouch::Any

csrrowptr::Any

csrcolval::Any

csrnzval::Any

csccolptr::Any

cscrowval::Any

cscnzval::Any

Assembler for dynamic problems without damping

Provides both a mass and stiffness matrix

FiniteElementContainers.DynamicAssembler — Method

Default initialization Assumes no dirichlet bcs

TODO add typing to constructor

FiniteElementContainers.StaticAssembler — Type

Is::AbstractVector

Js::AbstractVector

unknown_dofs::AbstractVector

block_sizes::AbstractVector

block_offsets::AbstractVector

residuals::AbstractArray

stiffnesses::AbstractVector

klasttouch::Any

csrrowptr::Any

csrcolval::Any

csrnzval::Any

csccolptr::Any

cscrowval::Any

cscnzval::Any

Assembler for static or quasistatic problems where only a stiffness matrix is necessary

FiniteElementContainers.StaticAssembler — Method

Default initialization Assumes no dirichlet bcs

TODO add typing to constructor

FiniteElementContainers.assemble! — Method

assemble!(
    assembler::DynamicAssembler,
    dof::DofManager,
    fspace::FunctionSpace,
    X,
    U,
    block_id,
    residual_func,
    tangent_func,
    mass_func
)

Simple method for assembling in serial

assemble!(
    assembler,
    dof,
    fspace,
    X,
    U,
    block_id,
    residual_func,
    tangent_func,
    mass_func
)

defined at /home/runner/work/FiniteElementContainers.jl/FiniteElementContainers.jl/src/assemblers/DynamicAssemblers.jl:162.

FiniteElementContainers.assemble! — Method

assemble!(
    assembler::StaticAssembler,
    dof::DofManager,
    fspace::FunctionSpace,
    X,
    U,
    block_id,
    residual_func,
    tangent_func
)

Simple method for assembling in serial

assemble!(
    assembler,
    dof,
    fspace,
    X,
    U,
    block_id,
    residual_func,
    tangent_func
)

defined at /home/runner/work/FiniteElementContainers.jl/FiniteElementContainers.jl/src/assemblers/StaticAssembler.jl:148.

FiniteElementContainers.assemble! — Method

assemble!(
    assembler::DynamicAssembler,
    K_el::AbstractMatrix{<:Number},
    M_el::AbstractMatrix{<:Number},
    block_id::Int64,
    el_id::Int64
) -> AbstractMatrix{<:Number}

assembly for stiffness matrix

assemble!(assembler, K_el, M_el, block_id, el_id)

defined at /home/runner/work/FiniteElementContainers.jl/FiniteElementContainers.jl/src/assemblers/DynamicAssemblers.jl:143.

FiniteElementContainers.assemble! — Method

assemble!(
    assembler::StaticAssembler,
    K_el::AbstractMatrix,
    block_id::Int64,
    el_id::Int64
) -> AbstractMatrix

assembly for stiffness matrix

assemble!(assembler, K_el, block_id, el_id)

defined at /home/runner/work/FiniteElementContainers.jl/FiniteElementContainers.jl/src/assemblers/StaticAssembler.jl:130.

FiniteElementContainers.assemble! — Method

assemble!(
    assembler::Assembler,
    R_el::AbstractVector{<:Number},
    conn::AbstractVector{<:Integer}
)

assembly method for just a residual vector

TODO need to add an Atomix lock here TODO add block_id to fspace or something like that

assemble!(assembler, R_el, conn)

defined at /home/runner/work/FiniteElementContainers.jl/FiniteElementContainers.jl/src/assemblers/Assemblers.jl:20.

FiniteElementContainers.bc_ids — Method

bc_ids(dof_manager::DofManager, nsets, dofs) -> Any

eventually use this method

bc_ids(dof_manager, nsets, dofs)

defined at /home/runner/work/FiniteElementContainers.jl/FiniteElementContainers.jl/src/DofManagers.jl:69.

FiniteElementContainers.file_name — Method

file_name(mesh::FiniteElementContainers.AbstractMesh) -> Any

Returns file name for an mesh type

file_name(mesh)

defined at /home/runner/work/FiniteElementContainers.jl/FiniteElementContainers.jl/src/Meshes.jl:18.

FiniteElementContainers.float_type — Method

float_type(_::Assembler{R, I}) -> Any

float_type(_)

defined at /home/runner/work/FiniteElementContainers.jl/FiniteElementContainers.jl/src/assemblers/Assemblers.jl:11.

FiniteElementContainers.int_type — Method

int_type(_::Assembler{R, I}) -> Any

int_type(_)

defined at /home/runner/work/FiniteElementContainers.jl/FiniteElementContainers.jl/src/assemblers/Assemblers.jl:7.

FiniteElementContainers.update_unknown_dofs! — Method

update_unknown_dofs!(
    dof::DofManager{T, ND, NN, A, V}
) -> Any

"Default" method to reset all dofs to unknown

update_unknown_dofs!(dof)

defined at /home/runner/work/FiniteElementContainers.jl/FiniteElementContainers.jl/src/DofManagers.jl:46.

FiniteElementContainers.update_unknown_dofs! — Method

update_unknown_dofs!(
    assembler::Assembler,
    dof,
    fspaces,
    nodes_in::AbstractVector{<:Integer}
) -> Any

method that assumes first dof TODO move sorting of nodes up stream TODO remove other scratch unknowns and unknown_dofs arrays

update_unknown_dofs!(assembler, dof, fspaces, nodes_in)

defined at /home/runner/work/FiniteElementContainers.jl/FiniteElementContainers.jl/src/assemblers/Assemblers.jl:35.

FiniteElementContainers.update_unknown_dofs! — Method

update_unknown_dofs!(
    dof_manager::DofManager,
    dofs::AbstractVector{<:Integer}
) -> Any

Method when all dofs are updated at once

First it resets all dofs to unknowns, then one by one sets dofs to bcs in dofs

Assumes there's a unique set of nodes provided

update_unknown_dofs!(dof_manager, dofs)

defined at /home/runner/work/FiniteElementContainers.jl/FiniteElementContainers.jl/src/DofManagers.jl:58.

SparseArrays.sparse! — Method

sparse!(
    assembler::DynamicAssembler
) -> Tuple{SparseArrays.SparseMatrixCSC, SparseArrays.SparseMatrixCSC}

sparse!(assembler)

defined at /home/runner/work/FiniteElementContainers.jl/FiniteElementContainers.jl/src/assemblers/DynamicAssemblers.jl:123.

SparseArrays.sparse! — Method

sparse!(
    assembler::StaticAssembler
) -> SparseArrays.SparseMatrixCSC

sparse!(assembler)

defined at /home/runner/work/FiniteElementContainers.jl/FiniteElementContainers.jl/src/assemblers/StaticAssembler.jl:117.

SparseArrays.sparse — Method

sparse(
    assembler::DynamicAssembler
) -> Tuple{SparseArrays.SparseMatrixCSC, SparseArrays.SparseMatrixCSC}

sparse(assembler)

defined at /home/runner/work/FiniteElementContainers.jl/FiniteElementContainers.jl/src/assemblers/DynamicAssemblers.jl:114.

SparseArrays.sparse — Method

sparse(
    assembler::StaticAssembler
) -> SparseArrays.SparseMatrixCSC

sparse(assembler)

defined at /home/runner/work/FiniteElementContainers.jl/FiniteElementContainers.jl/src/assemblers/StaticAssembler.jl:110.