FiniteElementContainers

Documentation for FiniteElementContainers.

FiniteElementContainers.FileMesh

FiniteElementContainers.Mesh

FiniteElementContainers.Mesh

FiniteElementContainers.file_name

FiniteElementContainers.FileMesh — Type

file_name::String

mesh_obj::Any

Mesh type that has a handle to an open mesh file object. This type's methods are "overridden" in extensions.

See FiniteElementContainersExodusExt for an example.

FiniteElementContainers.Mesh — Type

file_name::String

coords::NodalField

block_ids::AbstractVector{<:Integer}

conns::NamedTuple

elem_types::NamedTuple

nset_ids::AbstractVector{<:Integer}

nset_nodes::AbstractVector{<:AbstractVector{<:Integer}}

sset_ids::AbstractVector{<:Integer}

sset_elems::AbstractVector{<:AbstractVector{<:Integer}}

sset_sides::AbstractVector{<:AbstractVector{<:Integer}}

Mesh type that should have most everything one might want.

TODO probably need some type of face/edge sets as well TODO make some of these things optionals

FiniteElementContainers.Mesh — Method

Mesh constructor

FiniteElementContainers.file_name — Method

file_name(mesh::FiniteElementContainers.AbstractMesh) -> Any

Returns file name for an mesh type

file_name(mesh)

defined at /home/runner/work/FiniteElementContainers.jl/FiniteElementContainers.jl/src/Meshes.jl:18.