Assemblers

FiniteElementContainers.assemble! — Method

assemble!(
    global_val::H1Field,
    local_val,
    fspace,
    dof_conn,
    e,
    block_num
)

Assembly method for residuals TODO this will likely break on mixed space problems

FiniteElementContainers.assemble! — Method

assemble!(
    global_val::Vector,
    local_val,
    fspace,
    dof_conn,
    e,
    block_num
)

Assembly method for a scalar field stored as a size 1 vector

FiniteElementContainers.assemble! — Method

assemble!(
    R::AbstractArray{<:Number},
    R_el::AbstractArray{<:Number},
    conn::AbstractArray{<:Integer}
)

generic assembly method that directly goes into a vector

SparseArrays.sparse! — Method

sparse!(assembler::SparseMatrixAssembler) -> SparseMatrixCSC

SparseArrays.sparse — Method

sparse(
    assembler::SparseMatrixAssembler
) -> SparseMatrixCSC{Tv} where Tv<:Number