Assemblers

This section describes the assemblers that are currently available and their abstract interface.

All assemblers must possess at minimum a DofManager.

Matrices

FiniteElementContainers._assemble_block_matrix! — Method

_assemble_block_matrix!(
    field::AbstractVector{<:Number},
    pattern::FiniteElementContainers.SparsityPattern,
    physics::AbstractPhysics,
    ref_fe::ReferenceFiniteElements.ReferenceFE,
    U::FiniteElementContainers.AbstractField,
    X::FiniteElementContainers.AbstractField,
    state_old::L2QuadratureField,
    state_new::L2QuadratureField,
    props::Union{var"#s37", var"#s36"} where {var"#s37"<:(StaticArraysCore.SVector), var"#s36"<:L2ElementField},
    t::Number,
    dt::Number,
    conns::Connectivity,
    block_id::Int64,
    func::Function,
    backend::KernelAbstractions.Backend
)

Assembly method for a block labelled as block_id. This is a GPU agnostic implementation using KernelAbstractions and Atomix for eliminating race conditions

TODO add state variables and physics properties

FiniteElementContainers._assemble_block_matrix! — Method

_assemble_block_matrix!(
    field::AbstractVector{<:Number},
    pattern::FiniteElementContainers.SparsityPattern,
    physics::AbstractPhysics,
    ref_fe::ReferenceFiniteElements.ReferenceFE,
    U::FiniteElementContainers.AbstractField,
    X::FiniteElementContainers.AbstractField,
    state_old::L2QuadratureField,
    state_new::L2QuadratureField,
    props::Union{var"#s32", var"#s31"} where {var"#s32"<:(StaticArraysCore.SVector), var"#s31"<:L2ElementField},
    t::Number,
    dt::Number,
    conns::Connectivity,
    block_id::Int64,
    func::Function,
    _::KernelAbstractions.CPU
)

Assembly method for a block labelled as block_id. This is a CPU implementation with no threading.

TODO add state variables and physics properties

FiniteElementContainers.assemble_matrix! — Method

assemble_matrix!(
    assembler,
    Uu,
    p,
    _::Type{H1Field},
    func::Function
)

Note this is hard coded to storing the assembled sparse matrix in the stiffness_storage field of assembler.

Matrix Action

FiniteElementContainers._assemble_block_matrix_action! — Method

_assemble_block_matrix_action!(
    field::FiniteElementContainers.AbstractField,
    physics::AbstractPhysics,
    ref_fe::ReferenceFiniteElements.ReferenceFE,
    U::FiniteElementContainers.AbstractField,
    V::FiniteElementContainers.AbstractField,
    X::FiniteElementContainers.AbstractField,
    state_old::L2QuadratureField,
    state_new::L2QuadratureField,
    props::Union{var"#s35", var"#s34"} where {var"#s35"<:(StaticArraysCore.SVector), var"#s34"<:L2ElementField},
    t::Number,
    Δt::Number,
    conns::Connectivity,
    block_id::Int64,
    func::Function,
    _::KernelAbstractions.CPU
)

Assembly method for a block labelled as block_id. This is a CPU implementation with no threading.

TODO improve typing of fields to ensure they mathc up in terms of function spaces

FiniteElementContainers._assemble_block_matrix_action! — Method

_assemble_block_matrix_action!(
    field::FiniteElementContainers.AbstractField,
    physics::AbstractPhysics,
    ref_fe::ReferenceFiniteElements.ReferenceFE,
    U::FiniteElementContainers.AbstractField,
    V::FiniteElementContainers.AbstractField,
    X::FiniteElementContainers.AbstractField,
    state_old::L2QuadratureField,
    state_new::L2QuadratureField,
    props::Union{var"#s130", var"#s75"} where {var"#s130"<:(StaticArraysCore.SVector), var"#s75"<:L2ElementField},
    t::Number,
    Δt::Number,
    conns::Connectivity,
    block_id::Int64,
    func::Function,
    backend::KernelAbstractions.Backend
)

Assembly method for a block labelled as block_id. This is a GPU agnostic implementation using KernelAbstractions and Atomix for eliminating race conditions

TODO add state variables and physics properties

FiniteElementContainers._assemble_block_matrix_action_kernel! — Method

_assemble_block_matrix_action_kernel!(
    dev
) -> Union{KernelAbstractions.Kernel{KernelAbstractions.CPU, KernelAbstractions.NDIteration.DynamicSize, KernelAbstractions.NDIteration.DynamicSize, typeof(FiniteElementContainers.cpu__assemble_block_matrix_action_kernel!)}, KernelAbstractions.Kernel{Backend, KernelAbstractions.NDIteration.DynamicSize, KernelAbstractions.NDIteration.DynamicSize, typeof(FiniteElementContainers.gpu__assemble_block_matrix_action_kernel!)} where Backend<:KernelAbstractions.GPU}

Kernel for residual block assembly

TODO mark const fields

FiniteElementContainers._hvp — Method

_hvp(
    asm::FiniteElementContainers.AbstractAssembler,
    backend::KernelAbstractions.Backend
) -> Any

TODO note will only work for H1 spaces right now. TODO move me to Assemblers.jl

FiniteElementContainers._hvp — Method

_hvp(
    asm::FiniteElementContainers.AbstractAssembler,
    _::KernelAbstractions.CPU
) -> Any

Scalar

FiniteElementContainers._assemble_block_scalar! — Method

_assemble_block_scalar!(
    field::AbstractMatrix{<:Number},
    physics::AbstractPhysics,
    ref_fe::ReferenceFiniteElements.ReferenceFE,
    U::FiniteElementContainers.AbstractField,
    X::FiniteElementContainers.AbstractField,
    state_old::L2QuadratureField,
    state_new::L2QuadratureField,
    props::Union{var"#s130", var"#s75"} where {var"#s130"<:(StaticArraysCore.SVector), var"#s75"<:L2ElementField},
    t::Number,
    Δt::Number,
    conns::Connectivity,
    block_id::Int64,
    func::Function,
    _::KernelAbstractions.CPU
)

Assembly method for a block labelled as block_id. This is a CPU implementation with no threading.

TODO improve typing of fields to ensure they mathc up in terms of function spaces

FiniteElementContainers._assemble_block_scalar! — Method

_assemble_block_scalar!(
    field::AbstractMatrix{<:Number},
    physics::AbstractPhysics,
    ref_fe::ReferenceFiniteElements.ReferenceFE,
    U::FiniteElementContainers.AbstractField,
    X::FiniteElementContainers.AbstractField,
    state_old::L2QuadratureField,
    state_new::L2QuadratureField,
    props::Union{var"#s140", var"#s139"} where {var"#s140"<:(StaticArraysCore.SVector), var"#s139"<:L2ElementField},
    t::Number,
    Δt::Number,
    conns::Connectivity,
    block_id::Int64,
    func::Function,
    backend::KernelAbstractions.Backend
)

Assembly method for a block labelled as block_id. This is a GPU agnostic implementation using KernelAbstractions and Atomix for eliminating race conditions

TODO add state variables and physics properties

FiniteElementContainers._assemble_block_scalar_kernel! — Method

_assemble_block_scalar_kernel!(
    dev
) -> Union{KernelAbstractions.Kernel{KernelAbstractions.CPU, KernelAbstractions.NDIteration.DynamicSize, KernelAbstractions.NDIteration.DynamicSize, typeof(FiniteElementContainers.cpu__assemble_block_scalar_kernel!)}, KernelAbstractions.Kernel{Backend, KernelAbstractions.NDIteration.DynamicSize, KernelAbstractions.NDIteration.DynamicSize, typeof(FiniteElementContainers.gpu__assemble_block_scalar_kernel!)} where Backend<:KernelAbstractions.GPU}

Kernel for residual block assembly

TODO mark const fields

Vector

FiniteElementContainers._assemble_block_matrix_and_vector! — Method

_assemble_block_matrix_and_vector!(
    residual_field::FiniteElementContainers.AbstractField,
    stiffness_field::AbstractVector{<:Number},
    pattern::FiniteElementContainers.SparsityPattern,
    physics::AbstractPhysics,
    ref_fe::ReferenceFiniteElements.ReferenceFE,
    U::FiniteElementContainers.AbstractField,
    X::FiniteElementContainers.AbstractField,
    state_old::L2QuadratureField,
    state_new::L2QuadratureField,
    props::Union{var"#s32", var"#s31"} where {var"#s32"<:(StaticArraysCore.SVector), var"#s31"<:L2ElementField},
    t::Number,
    dt::Number,
    conns::Connectivity,
    block_id::Int64,
    residual_func::Function,
    stiffness_func::Function,
    _::KernelAbstractions.CPU
)

Assembly method for a block labelled as block_id. This is a CPU implementation with no threading.

TODO add state variables and physics properties TODO remove Float64 typing below for eventual unitful use

FiniteElementContainers._assemble_block_vector! — Method

_assemble_block_vector!(
    field::FiniteElementContainers.AbstractField,
    physics::AbstractPhysics,
    ref_fe::ReferenceFiniteElements.ReferenceFE,
    U::FiniteElementContainers.AbstractField,
    X::FiniteElementContainers.AbstractField,
    state_old::L2QuadratureField,
    state_new::L2QuadratureField,
    props::Union{var"#s138", var"#s131"} where {var"#s138"<:(StaticArraysCore.SVector), var"#s131"<:L2ElementField},
    t::Number,
    Δt::Number,
    conns::Connectivity,
    block_id::Int64,
    func::Function,
    _::KernelAbstractions.CPU
)

Assembly method for a block labelled as block_id. This is a CPU implementation with no threading.

TODO improve typing of fields to ensure they mathc up in terms of function spaces

FiniteElementContainers._assemble_block_vector! — Method

_assemble_block_vector!(
    field::FiniteElementContainers.AbstractField,
    physics::AbstractPhysics,
    ref_fe::ReferenceFiniteElements.ReferenceFE,
    U::FiniteElementContainers.AbstractField,
    X::FiniteElementContainers.AbstractField,
    state_old::L2QuadratureField,
    state_new::L2QuadratureField,
    props::Union{var"#s142", var"#s141"} where {var"#s142"<:(StaticArraysCore.SVector), var"#s141"<:L2ElementField},
    t::Number,
    Δt::Number,
    conns::Connectivity,
    block_id::Int64,
    func::Function,
    backend::KernelAbstractions.Backend
)

Assembly method for a block labelled as block_id. This is a GPU agnostic implementation using KernelAbstractions and Atomix for eliminating race conditions

TODO add state variables and physics properties

FiniteElementContainers._assemble_block_vector_kernel! — Method

_assemble_block_vector_kernel!(
    dev
) -> Union{KernelAbstractions.Kernel{KernelAbstractions.CPU, KernelAbstractions.NDIteration.DynamicSize, KernelAbstractions.NDIteration.DynamicSize, typeof(FiniteElementContainers.cpu__assemble_block_vector_kernel!)}, KernelAbstractions.Kernel{Backend, KernelAbstractions.NDIteration.DynamicSize, KernelAbstractions.NDIteration.DynamicSize, typeof(FiniteElementContainers.gpu__assemble_block_vector_kernel!)} where Backend<:KernelAbstractions.GPU}

Kernel for residual block assembly

TODO mark const fields

FiniteElementContainers._extract_residual_unknowns! — Method

_extract_residual_unknowns!(
    dev
) -> Union{KernelAbstractions.Kernel{KernelAbstractions.CPU, KernelAbstractions.NDIteration.DynamicSize, KernelAbstractions.NDIteration.DynamicSize, typeof(FiniteElementContainers.cpu__extract_residual_unknowns!)}, KernelAbstractions.Kernel{Backend, KernelAbstractions.NDIteration.DynamicSize, KernelAbstractions.NDIteration.DynamicSize, typeof(FiniteElementContainers.gpu__extract_residual_unknowns!)} where Backend<:KernelAbstractions.GPU}

FiniteElementContainers._residual — Method

_residual(
    asm::FiniteElementContainers.AbstractAssembler,
    backend::KernelAbstractions.Backend
) -> Any

FiniteElementContainers._residual — Method

_residual(
    asm::FiniteElementContainers.AbstractAssembler,
    _::KernelAbstractions.CPU
) -> Any

FiniteElementContainers.assemble_vector! — Method

assemble_vector!(
    assembler,
    Uu,
    p,
    _::Type{H1Field},
    func::Function
)

Abstract Interface

FiniteElementContainers.AbstractAssembler — Type

abstract type AbstractAssembler{Dof<:DofManager}

FiniteElementContainers._assemble_element! — Method

_assemble_element!(
    global_val::H1Field,
    local_val,
    conn,
    e,
    b
)

Assembly method for an H1Field, e.g. internal force Called on a single element e for a given block b where local_val has already been constructed from quadrature contributions.

FiniteElementContainers._assemble_element! — Method

_assemble_element!(
    global_val::AbstractVector{<:Number},
    local_val,
    conn,
    e,
    b
)

Assembly method for a scalar field stored as a size 1 vector Called on a single element e for a given block b where local_val has already been constructed from quadrature contributions.

FiniteElementContainers._cell_interpolants — Method

_cell_interpolants(
    ref_fe::ReferenceFiniteElements.ReferenceFE,
    q::Int64
) -> Any

FiniteElementContainers._element_level_fields — Method

_element_level_fields(U::H1Field, ref_fe, conns, e) -> Any

FiniteElementContainers._element_level_properties — Method

_element_level_properties(
    props::L2ElementField,
    e::Int64
) -> Any

FiniteElementContainers._element_level_properties — Method

_element_level_properties(
    props::StaticArraysCore.SArray{Tuple{NP}, T, 1, NP},
    _::Int64
) -> StaticArraysCore.SVector

FiniteElementContainers._quadrature_level_state — Method

_quadrature_level_state(
    state::L2QuadratureField,
    q::Int64,
    e::Int64
) -> Any

FiniteElementContainers.assemble! — Method

assemble!(
    assembler,
    Uu,
    p,
    Vv,
    sym::Symbol,
    type::Type{H1Field}
)

Top level assembly method for H1Field that loops over blocks and dispatches to appropriate kernels based on sym.

TODO need to make sure at setup time that physics and elem_conns have the same values order. Otherwise, shenanigans.

TODO figure out how to do generated functions

creates one type instability from the Val

FiniteElementContainers.assemble! — Method

assemble!(
    assembler,
    Uu,
    p,
    sym::Symbol,
    type::Type{H1Field}
) -> Any

Top level assembly method for H1Field that loops over blocks and dispatches to appropriate kernels based on sym.

TODO need to make sure at setup time that physics and elem_conns have the same values order. Otherwise, shenanigans.

TODO figure out how to do generated functions

creates one type instability from the Val

FiniteElementContainers.create_field — Method

create_field(
    asm::FiniteElementContainers.AbstractAssembler,
    type
) -> Union{Nothing, H1Field{Float64, _A, Vector{Float64}} where _A}

FiniteElementContainers.create_unknowns — Method

create_unknowns(
    asm::FiniteElementContainers.AbstractAssembler,
    type::Type{<:FiniteElementContainers.AbstractField}
) -> Any

FiniteElementContainers.create_unknowns — Method

create_unknowns(
    asm::FiniteElementContainers.AbstractAssembler
) -> Any

FiniteElementContainers.function_space — Method

function_space(
    asm::FiniteElementContainers.AbstractAssembler,
    _::Type{<:H1Field}
) -> Any

Returns either nothing or the function space associated with H1Fields in the problem.

FiniteElementContainers.mass — Method

mass(
    assembler::FiniteElementContainers.AbstractAssembler
) -> SparseArrays.SparseMatrixCSC{Float64, Int64}

FiniteElementContainers.residual — Method

residual(
    asm::FiniteElementContainers.AbstractAssembler
) -> Any

FiniteElementContainers.stiffness — Method

stiffness(
    assembler::FiniteElementContainers.AbstractAssembler
) -> SparseArrays.SparseMatrixCSC{Float64, Int64}

KernelAbstractions.get_backend — Method

get_backend(
    asm::FiniteElementContainers.AbstractAssembler
) -> Any

SparseMatrixAssembler

FiniteElementContainers.SparseMatrixAssembler — Type

struct SparseMatrixAssembler{Dof<:DofManager, Pattern<:FiniteElementContainers.SparsityPattern, Storage1<:(AbstractVector{<:Number}), Storage2<:FiniteElementContainers.AbstractField, Storage3<:NamedTuple} <: FiniteElementContainers.AbstractAssembler{Dof<:DofManager}

General sparse matrix assembler that can handle first or second order problems in time.

FiniteElementContainers.SparseMatrixAssembler — Method

SparseMatrixAssembler(
    dof::DofManager,
    type::Type{<:H1Field}
) -> SparseMatrixAssembler{Dof, Pattern, Vector{Float64}, Storage2, Storage3} where {Dof<:DofManager, Pattern<:(FiniteElementContainers.SparsityPattern{Vector{Int64}, _A, Vector{Float64}} where _A), Storage2<:(H1Field{Float64, _A, Vector{Float64}} where _A), Storage3<:(NamedTuple{_A, Tuple{Vararg{Matrix{Float64}, N}}} where {_A, N})}

Construct a SparseMatrixAssembler for a specific field type, e.g. H1Field. Can be used to create block arrays for mixed FEM problems.

FiniteElementContainers._assemble_element! — Method

_assemble_element!(
    pattern::FiniteElementContainers.SparsityPattern,
    storage,
    K_el::StaticArraysCore.SMatrix,
    el_id::Int64,
    block_id::Int64
)

Specialization of of _assemble_element! for SparseMatrixAssembler.

SparseArrays.sparse! — Method

sparse!(
    assembler::SparseMatrixAssembler,
    _::Val{:mass}
) -> SparseArrays.SparseMatrixCSC{Float64, Int64}

TODO add symbol to interface

SparseArrays.sparse! — Method

sparse!(
    assembler::SparseMatrixAssembler,
    _::Val{:stiffness}
) -> SparseArrays.SparseMatrixCSC{Float64, Int64}

TODO add symbol to interface

SparsityPattern

FiniteElementContainers.SparsityPattern — Type

struct SparsityPattern{I<:AbstractVector{Int64}, B, R<:AbstractVector{Float64}}

Is::AbstractVector{Int64}

Js::AbstractVector{Int64}

unknown_dofs::AbstractVector{Int64}

block_sizes::Any

block_offsets::Any

klasttouch::AbstractVector{Int64}

csrrowptr::AbstractVector{Int64}

csrcolval::AbstractVector{Int64}

csrnzval::AbstractVector{Float64}

csccolptr::AbstractVector{Int64}

cscrowval::AbstractVector{Int64}

cscnzval::AbstractVector{Float64}

Book-keeping struct for sparse matrices in FEM settings. This has all the information to construct a sparse matrix for either case where you want to eliminate fixed-dofs or not.