Assemblers

This section describes the assemblers that are currently available and their abstract interface.

All assemblers must possess at minimum a DofManager.

Abstract Interface

FiniteElementContainers.AbstractAssembler — Type

abstract type AbstractAssembler{Dof<:DofManager}

FiniteElementContainers._assemble_element! — Method

_assemble_element!(
    global_val::H1Field,
    local_val,
    conn,
    e,
    b
)

Assembly method for an H1Field, e.g. internal force Called on a single element e for a given block b where local_val has already been constructed from quadrature contributions.

FiniteElementContainers._assemble_element! — Method

_assemble_element!(
    global_val::AbstractVector{<:Number},
    local_val,
    conn,
    e,
    b
)

Assembly method for a scalar field stored as a size 1 vector Called on a single element e for a given block b where local_val has already been constructed from quadrature contributions.

FiniteElementContainers._element_level_coordinates — Method

_element_level_coordinates(
    X::H1Field,
    ref_fe,
    conns,
    e
) -> Any

FiniteElementContainers._element_level_fields — Method

_element_level_fields(U::H1Field, ref_fe, conns, e) -> Any

FiniteElementContainers._zero_storage — Method

_zero_storage(
    asm::FiniteElementContainers.AbstractAssembler,
    _::Val{:residual}
) -> Any

FiniteElementContainers.assemble! — Method

assemble!(assembler, physics, U::H1Field, sym)

Top level assembly method for H1Field that loops over blocks and dispatches to appropriate kernels based on sym.

FiniteElementContainers.create_bcs — Method

create_bcs(
    asm::FiniteElementContainers.AbstractAssembler,
    type
) -> Any

FiniteElementContainers.create_field — Method

create_field(
    asm::FiniteElementContainers.AbstractAssembler,
    type
) -> Union{Nothing, H1Field{Float64, _A, Vector{Float64}} where _A}

FiniteElementContainers.create_unknowns — Method

create_unknowns(
    asm::FiniteElementContainers.AbstractAssembler,
    type::Type{<:FiniteElementContainers.AbstractField}
) -> Any

FiniteElementContainers.create_unknowns — Method

create_unknowns(
    asm::FiniteElementContainers.AbstractAssembler
) -> Any

FiniteElementContainers.residual — Method

residual(
    asm::FiniteElementContainers.AbstractAssembler
) -> Any

FiniteElementContainers.update_field! — Method

update_field!(
    U,
    asm::FiniteElementContainers.AbstractAssembler,
    Uu,
    Ubc
)

KernelAbstractions.get_backend — Method

get_backend(
    asm::FiniteElementContainers.AbstractAssembler
) -> Any

Abstract Interface - CPU Specialization

FiniteElementContainers._assemble_block_residual! — Method

_assemble_block_residual!(
    assembler,
    physics,
    ref_fe,
    U,
    X,
    conns,
    block_id,
    _::KernelAbstractions.CPU
)

Assembly method for a block labelled as block_id. This is a CPU implementation with no threading.

TODO add state variables and physics properties TODO remove Float64 typing below for eventual unitful use

FiniteElementContainers._residual — Method

_residual(
    asm::FiniteElementContainers.AbstractAssembler,
    _::KernelAbstractions.CPU
) -> Any

Abstract Interface - GPU Specialization

FiniteElementContainers._assemble_block_residual! — Method

_assemble_block_residual!(
    assembler,
    physics,
    ref_fe,
    U,
    X,
    conns,
    block_id,
    backend::KernelAbstractions.Backend
)

Assembly method for a block labelled as block_id. This is a GPU agnostic implementation using KernelAbstractions and Atomix for eliminating race conditions

TODO add state variables and physics properties

FiniteElementContainers._assemble_block_residual_kernel! — Method

_assemble_block_residual_kernel!(
    dev
) -> Union{KernelAbstractions.Kernel{KernelAbstractions.CPU, KernelAbstractions.NDIteration.DynamicSize, KernelAbstractions.NDIteration.DynamicSize, typeof(FiniteElementContainers.cpu__assemble_block_residual_kernel!)}, KernelAbstractions.Kernel{Backend, KernelAbstractions.NDIteration.DynamicSize, KernelAbstractions.NDIteration.DynamicSize, typeof(FiniteElementContainers.gpu__assemble_block_residual_kernel!)} where Backend<:KernelAbstractions.GPU}

Kernel for residual block assembly

FiniteElementContainers._assemble_block_stiffness! — Method

_assemble_block_stiffness!(
    assembler,
    physics,
    ref_fe,
    U,
    X,
    conns,
    block_id,
    backend::KernelAbstractions.Backend
)

Assembly method for a block labelled as block_id. This is a GPU agnostic implementation using KernelAbstractions and Atomix for eliminating race conditions

TODO add state variables and physics properties

FiniteElementContainers._extract_residual_unknowns! — Method

_extract_residual_unknowns!(
    dev
) -> Union{KernelAbstractions.Kernel{KernelAbstractions.CPU, KernelAbstractions.NDIteration.DynamicSize, KernelAbstractions.NDIteration.DynamicSize, typeof(FiniteElementContainers.cpu__extract_residual_unknowns!)}, KernelAbstractions.Kernel{Backend, KernelAbstractions.NDIteration.DynamicSize, KernelAbstractions.NDIteration.DynamicSize, typeof(FiniteElementContainers.gpu__extract_residual_unknowns!)} where Backend<:KernelAbstractions.GPU}

FiniteElementContainers._residual — Method

_residual(
    asm::FiniteElementContainers.AbstractAssembler,
    backend::KernelAbstractions.Backend
) -> Any

SparsityPattern

FiniteElementContainers.SparsityPattern — Type

struct SparsityPattern{I<:AbstractVector{Int64}, B, R<:AbstractVector{Float64}}

Is::AbstractVector{Int64}

Js::AbstractVector{Int64}

unknown_dofs::AbstractVector{Int64}

block_sizes::Any

block_offsets::Any

klasttouch::AbstractVector{Int64}

csrrowptr::AbstractVector{Int64}

csrcolval::AbstractVector{Int64}

csrnzval::AbstractVector{Float64}

csccolptr::AbstractVector{Int64}

cscrowval::AbstractVector{Int64}

cscnzval::AbstractVector{Float64}

Book-keeping struct for sparse matrices in FEM settings. This has all the information to construct a sparse matrix for either case where you want to eliminate fixed-dofs or not.

SparseMatrixAssembler

FiniteElementContainers.SparseMatrixAssembler — Type

struct SparseMatrixAssembler{Dof<:DofManager, Pattern<:FiniteElementContainers.SparsityPattern, Storage1<:(AbstractArray{<:Number}), Storage2<:FiniteElementContainers.AbstractField, Storage3<:(AbstractVector{<:Number})} <: FiniteElementContainers.AbstractAssembler{Dof<:DofManager}

General sparse matrix assembler that can handle first or second order problems in time.

FiniteElementContainers.SparseMatrixAssembler — Method

SparseMatrixAssembler(
    dof::DofManager,
    type::Type{<:H1Field}
) -> SparseMatrixAssembler{Dof, Pattern, Vector{Float64}, Storage2} where {Dof<:DofManager, Pattern<:(FiniteElementContainers.SparsityPattern{Vector{Int64}, _A, Vector{Float64}} where _A), Storage2<:(H1Field{Float64, _A, Vector{Float64}} where _A)}

Construct a SparseMatrixAssembler for a specific field type, e.g. H1Field. Can be used to create block arrays for mixed FEM problems.

FiniteElementContainers._assemble_block_mass! — Method

_assemble_block_mass!(
    assembler,
    physics,
    ref_fe,
    U,
    X,
    conns,
    block_id,
    _::KernelAbstractions.CPU
)

Assembly method for a block labelled as block_id. This is a CPU implementation with no threading.

TODO add state variables and physics properties

FiniteElementContainers._assemble_block_residual_and_stiffness! — Method

_assemble_block_residual_and_stiffness!(
    assembler,
    physics,
    ref_fe,
    U,
    X,
    conns,
    block_id,
    _::KernelAbstractions.CPU
)

Assembly method for a block labelled as block_id. This is a CPU implementation with no threading.

TODO add state variables and physics properties TODO remove Float64 typing below for eventual unitful use

FiniteElementContainers._assemble_block_stiffness! — Method

_assemble_block_stiffness!(
    assembler,
    physics,
    ref_fe,
    U,
    X,
    conns,
    block_id,
    _::KernelAbstractions.CPU
)

Assembly method for a block labelled as block_id. This is a CPU implementation with no threading.

TODO add state variables and physics properties

FiniteElementContainers._assemble_element! — Method

_assemble_element!(
    asm::SparseMatrixAssembler,
    sym,
    K_el::StaticArraysCore.SMatrix,
    conn,
    el_id::Int64,
    block_id::Int64
)

Specialization of of _assemble_element! for SparseMatrixAssembler.

SparseArrays.sparse! — Method

sparse!(
    assembler::SparseMatrixAssembler,
    sym
) -> SparseArrays.SparseMatrixCSC{Float64, Int64}

TODO add symbol to interface

SparseArrays.sparse — Method

sparse(
    assembler::SparseMatrixAssembler
) -> SparseArrays.SparseMatrixCSC{Tv, Int64} where Tv<:Number