Assemblers

FiniteElementContainers.Assembler — Type

abstract type Assembler{Rtype, Itype}

FiniteElementContainers.assemble! — Method

assemble!(
    assembler::Assembler,
    R_el::AbstractVector{<:Number},
    conn::AbstractVector{<:Integer}
)

assembly method for just a residual vector

TODO need to add an Atomix lock here TODO add block_id to fspace or something like that

FiniteElementContainers.float_type — Method

float_type(_::Assembler{R, I}) -> Any

FiniteElementContainers.int_type — Method

int_type(_::Assembler{R, I}) -> Any

FiniteElementContainers.update_unknown_dofs! — Method

update_unknown_dofs!(
    assembler::Assembler,
    dof,
    fspaces,
    nodes_in::AbstractVector{<:Integer}
) -> Any

method that assumes first dof TODO move sorting of nodes up stream TODO remove other scratch unknowns and unknown_dofs arrays

FiniteElementContainers.DynamicAssembler — Type

struct DynamicAssembler{Rtype, Itype, I<:AbstractArray{Itype, 1}, J<:AbstractArray{Itype, 1}, U<:AbstractArray{Itype, 1}, Sizes<:AbstractArray{Itype, 1}, Offsets<:AbstractArray{Itype, 1}, R<:(AbstractArray{Rtype}), K<:AbstractArray{Rtype, 1}, M<:AbstractArray{Rtype, 1}, C1, C2, C3, C4, C5, C6, C7} <: Assembler{Rtype, Itype}

Is::AbstractVector

Js::AbstractVector

unknown_dofs::AbstractVector

block_sizes::AbstractVector

block_offsets::AbstractVector

residuals::AbstractArray

stiffnesses::AbstractVector

masses::AbstractVector

klasttouch::Any

csrrowptr::Any

csrcolval::Any

csrnzval::Any

csccolptr::Any

cscrowval::Any

cscnzval::Any

Assembler for dynamic problems without damping

Provides both a mass and stiffness matrix

FiniteElementContainers.StaticAssembler — Type

struct StaticAssembler{Rtype, Itype, I<:AbstractArray{Itype, 1}, J<:AbstractArray{Itype, 1}, U<:AbstractArray{Itype, 1}, Sizes<:AbstractArray{Itype, 1}, Offsets<:AbstractArray{Itype, 1}, R<:(AbstractArray{Rtype}), K<:AbstractArray{Rtype, 1}, C1, C2, C3, C4, C5, C6, C7} <: Assembler{Rtype, Itype}

Is::AbstractVector

Js::AbstractVector

unknown_dofs::AbstractVector

block_sizes::AbstractVector

block_offsets::AbstractVector

residuals::AbstractArray

stiffnesses::AbstractVector

klasttouch::Any

csrrowptr::Any

csrcolval::Any

csrnzval::Any

csccolptr::Any

cscrowval::Any

cscnzval::Any

Assembler for static or quasistatic problems where only a stiffness matrix is necessary

FiniteElementContainers.assemble! — Function

assemble!(
    assembler::Assembler,
    R_el::AbstractVector{<:Number},
    conn::AbstractVector{<:Integer}
)

assembly method for just a residual vector

TODO need to add an Atomix lock here TODO add block_id to fspace or something like that

assemble!(
    assembler::DynamicAssembler,
    K_el::AbstractMatrix{<:Number},
    M_el::AbstractMatrix{<:Number},
    block_id::Int64,
    el_id::Int64
) -> AbstractMatrix{<:Number}

assembly for stiffness matrix

assemble!(
    assembler::DynamicAssembler,
    dof::DofManager,
    fspace::FunctionSpace,
    X,
    U,
    block_id,
    residual_func,
    tangent_func,
    mass_func
)

Simple method for assembling in serial

assemble!(
    assembler::DynamicAssembler,
    dof::DofManager,
    fspaces,
    X,
    U,
    residual_func,
    tangent_func,
    mass_func
)

assemble!(
    assembler::StaticAssembler,
    K_el::AbstractMatrix,
    block_id::Int64,
    el_id::Int64
) -> AbstractMatrix

assembly for stiffness matrix

assemble!(
    assembler::StaticAssembler,
    dof::DofManager,
    fspace::FunctionSpace,
    X,
    U,
    block_id,
    residual_func,
    tangent_func
)

Simple method for assembling in serial

assemble!(
    assembler::StaticAssembler,
    dof::DofManager,
    fspaces,
    X,
    U,
    residual_func,
    tangent_func
)

Top level method using methods