Fields

Fields serve as loose wrappers around AbstractArray subtypes such that the size of array slices are known at compile time. Although this introduces a type-instability, the idea is to do this at the top most level (mainly at setup time of a FEM simulation). By introducing this type instability, we can gain information about the field type that is used in methods downstream to construct StaticArrays of views of field types.

All fields are subtypes of the abstract type AbstractField

julia> using FiniteElementContainers
julia> FiniteElementContainers.AbstractFieldFiniteElementContainers.AbstractField

Example - H1Field a.k.a. NodalField

We can set up a H1Field in one of two ways. The simplest constructor form can be used as follows

julia> using FiniteElementContainers
julia> field = H1Field(rand(2, 10))2×10 H1Field{Float64, Vector{Float64}, 2}:
 0.231577  0.378486  0.961391  0.731024  …  0.16172  0.270843  0.252361
 0.997933  0.372066  0.501172  0.216361     0.59497  0.797727  0.235899

This is stored in a vectorized way as can be seen above

julia> field.data20-element Vector{Float64}:
 0.23157737448764337
 0.9979333536218965
 0.3784863094031573
 0.3720662009298886
 0.9613913716332062
 0.5011715927309038
 0.7310244166593413
 0.21636074143355455
 0.7622427797372786
 0.6493199701247251
 0.7173305000281625
 0.09944322922945292
 0.9993555422212509
 0.2314675045352813
 0.16172022054975288
 0.5949699684601365
 0.2708429398626083
 0.7977266858520533
 0.2523613356473927
 0.23589948291907736

Fields can be indexed like regular arrays, e.g.

julia> field[1, 1]0.23157737448764337

julia> field[1, :]10-element Vector{Float64}:
 0.23157737448764337
 0.3784863094031573
 0.9613913716332062
 0.7310244166593413
 0.7622427797372786
 0.7173305000281625
 0.9993555422212509
 0.16172022054975288
 0.2708429398626083
 0.2523613356473927

etc.

Abstract type

The base type for fields is the AbstractField abstract type.

FiniteElementContainers.AbstractField — Type

abstract type AbstractField{T, N, D<:AbstractArray{T, 1}} <: AbstractArray{T, N}

Thin wrapper that subtypes AbstractArray and serves as the base Field type

Any new field added to FiniteElementContainers should be a subtype of this type.

Methods for AbstractField

Base.fill! — Method

fill!(
    field::FiniteElementContainers.AbstractField{T, N, D},
    v
) -> Any

Base.unique — Method

unique(field::FiniteElementContainers.AbstractField) -> Any

FiniteElementContainers.num_entities — Method

num_entities(
    field::FiniteElementContainers.AbstractContinuousField{T, D, NF}
) -> Any

FiniteElementContainers.num_fields — Method

num_fields(
    _::FiniteElementContainers.AbstractContinuousField{T, D, NF}
) -> Any

KernelAbstractions.get_backend — Method

get_backend(
    field::FiniteElementContainers.AbstractField
) -> Any

Implementations

The existing direct subtypes of AbstractField are the following

Connectivity

The connectivity type is a simple alias for L2ElementField defined below

FiniteElementContainers.Connectivity — Type

struct Connectivity{T<:Integer, D<:AbstractArray{T<:Integer, 1}}

H1 field

FiniteElementContainers.H1Field — Type

struct H1Field{T, D, NF} <: FiniteElementContainers.AbstractContinuousField{T, D, NF}

Implementation of fields that live on nodes.

FiniteElementContainers.H1Field — Method

H1Field(data::AbstractMatrix) -> H1Field

Hcurl field

FiniteElementContainers.HcurlField — Type

struct HcurlField{T, D, NF} <: FiniteElementContainers.AbstractContinuousField{T, D, NF}

Implementation of fields that live in Hdiv spaces.

FiniteElementContainers.HcurlField — Method

HcurlField(data::AbstractMatrix) -> HcurlField

Hdiv field

FiniteElementContainers.HdivField — Type

struct HdivField{T, D, NF} <: FiniteElementContainers.AbstractContinuousField{T, D, NF}

Implementation of fields that live in Hdiv spaces.

FiniteElementContainers.HdivField — Method

HdivField(data::AbstractMatrix) -> HdivField

L2 field