Meshes

FiniteElementContainers.AbstractMesh — Type

abstract type AbstractMesh <: FiniteElementContainers.FEMContainer

FiniteElementContainers.FileMesh — Type

struct FileMesh{MeshObj} <: FiniteElementContainers.AbstractMesh

file_name::String

mesh_obj::Any

Mesh type that has a handle to an open mesh file object. This type's methods are "overridden" in extensions.

See FiniteElementContainersExodusExt for an example.

FiniteElementContainers.coordinates — Function

Dummy method to be overriden for specific mesh file format

FiniteElementContainers.copy_mesh — Function

Dummy method to be overriden for specific mesh file format

FiniteElementContainers.element_block_ids — Function

Dummy method to be overriden for specific mesh file format

FiniteElementContainers.element_connectivity — Function

Dummy method to be overriden for specific mesh file format

FiniteElementContainers.element_type — Function

Dummy method to be overriden for specific mesh file format

FiniteElementContainers.file_name — Method

file_name(mesh::FiniteElementContainers.AbstractMesh) -> Any

Returns file name for an mesh type

FiniteElementContainers.nodeset — Function

Dummy method to be overriden for specific mesh file format

FiniteElementContainers.nodeset_ids — Function

Dummy method to be overriden for specific mesh file format

FiniteElementContainers.num_dimensions — Function

Dummy method to be overriden for specific mesh file format

FiniteElementContainers.num_nodes — Function

Dummy method to be overriden for specific mesh file format

FiniteElementContainers.sideset — Function

Dummy method to be overriden for specific mesh file format

FiniteElementContainers.sideset_ids — Function

Dummy method to be overriden for specific mesh file format