Meshes

FiniteElementContainers.AbstractMesh — Type

abstract type AbstractMesh <: FiniteElementContainers.FEMContainer

FiniteElementContainers.FileMesh — Type

struct FileMesh{MeshObj} <: FiniteElementContainers.AbstractMesh

file_name::String

mesh_obj::Any

Mesh type that has a handle to an open mesh file object. This type's methods are "overridden" in extensions.

See FiniteElementContainersExodusExt for an example.

FiniteElementContainers.UnstructuredMesh — Type

struct UnstructuredMesh{MeshObj, X, EBlockNames, ETypes, EConns, EMaps, NSetNodes, SSetElems, SSetNodes, SSetSides, EdgeConns, FaceConns} <: FiniteElementContainers.AbstractMesh

mesh_obj::Any

nodal_coords::Any

element_block_names::Any

element_types::Any

element_conns::Any

element_id_maps::Any

nodeset_nodes::Any

sideset_elems::Any

sideset_nodes::Any

sideset_sides::Any

edge_conns::Any

face_conns::Any

FiniteElementContainers.UnstructuredMesh — Method

UnstructuredMesh(
    file_type,
    file_name::String,
    create_edges::Bool,
    create_faces::Bool
) -> UnstructuredMesh{MeshObj, X, EBlockNames, ETypes, EConns, EMaps, NSetNodes, SSetElems, SSetNodes, SSetSides, EdgeConns, Nothing} where {MeshObj, X, EBlockNames, ETypes, EConns, EMaps, NSetNodes, SSetElems, SSetNodes, SSetSides, EdgeConns}

FiniteElementContainers.UnstructuredMesh — Method

UnstructuredMesh(
    file_name::String;
    create_edges,
    create_faces
) -> UnstructuredMesh{MeshObj, X, EBlockNames, ETypes, EConns, EMaps, NSetNodes, SSetElems, SSetNodes, SSetSides, EdgeConns, Nothing} where {MeshObj, X, EBlockNames, ETypes, EConns, EMaps, NSetNodes, SSetElems, SSetNodes, SSetSides, EdgeConns}

FiniteElementContainers.coordinates — Method

coordinates(
    _::FiniteElementContainers.AbstractMesh
) -> Matrix

Dummy method to be overriden for specific mesh file format

FiniteElementContainers.copy_mesh — Function

Dummy method to be overriden for specific mesh file format

FiniteElementContainers.element_block_id_map — Method

element_block_id_map(
    _::FiniteElementContainers.AbstractMesh,
    id
) -> Any

Dummy method to be overriden for specific mesh file format

FiniteElementContainers.element_block_ids — Method

element_block_ids(
    _::FiniteElementContainers.AbstractMesh
) -> Any

Dummy method to be overriden for specific mesh file format

FiniteElementContainers.element_block_names — Method

element_block_names(
    _::FiniteElementContainers.AbstractMesh
) -> Any

Dummy method to be overriden for specific mesh file format

FiniteElementContainers.element_connectivity — Method

element_connectivity(
    _::FiniteElementContainers.AbstractMesh
)

Dummy method to be overriden for specific mesh file format

FiniteElementContainers.element_type — Method

element_type(_::FiniteElementContainers.AbstractMesh)

Dummy method to be overriden for specific mesh file format

FiniteElementContainers.file_name — Method

file_name(mesh::FiniteElementContainers.AbstractMesh) -> Any

Returns file name for an mesh type

FiniteElementContainers.nodeset — Method

nodeset(_::FiniteElementContainers.AbstractMesh)

Dummy method to be overriden for specific mesh file format

FiniteElementContainers.nodeset_ids — Method

nodeset_ids(_::FiniteElementContainers.AbstractMesh) -> Any

Dummy method to be overriden for specific mesh file format

FiniteElementContainers.nodeset_names — Method

nodeset_names(
    _::FiniteElementContainers.AbstractMesh
) -> Any

Dummy method to be overriden for specific mesh file format

FiniteElementContainers.nodesets — Method

nodesets(
    _::FiniteElementContainers.AbstractMesh,
    ids
) -> Any

Dummy method to be overriden for specific mesh file format

FiniteElementContainers.num_dimensions — Method

num_dimensions(
    _::FiniteElementContainers.AbstractMesh
) -> Int32

Dummy method to be overriden for specific mesh file format

FiniteElementContainers.num_nodes — Method

num_nodes(_::FiniteElementContainers.AbstractMesh) -> Int32

Dummy method to be overriden for specific mesh file format

FiniteElementContainers.sideset — Method

sideset(_::FiniteElementContainers.AbstractMesh)

Dummy method to be overriden for specific mesh file format

FiniteElementContainers.sideset_ids — Method

sideset_ids(_::FiniteElementContainers.AbstractMesh) -> Any

Dummy method to be overriden for specific mesh file format

FiniteElementContainers.sideset_names — Method

sideset_names(
    _::FiniteElementContainers.AbstractMesh
) -> Any

Dummy method to be overriden for specific mesh file format

FiniteElementContainers.sidesets — Method

sidesets(
    _::FiniteElementContainers.AbstractMesh,
    ids
) -> Any

Dummy method to be overriden for specific mesh file format