Parameters

FiniteElementContainers.AbstractParameters — Type

abstract type AbstractParameters

FiniteElementContainers.Parameters — Type

struct Parameters{RT<:Number, RV<:AbstractArray{RT<:Number, 1}, IV<:(AbstractVector{<:Integer}), ICFuncs<:NamedTuple, DBCFuncs<:NamedTuple, NBCs<:NeumannBCs, Phys<:NamedTuple, Props<:NamedTuple, S<:NamedTuple, NDims, NH1Fields} <: FiniteElementContainers.AbstractParameters

ics::InitialConditions{IV, RV} where {RT<:Number, RV<:AbstractVector{RT}, IV<:(AbstractVector{<:Integer})}

dirichlet_bcs::DirichletBCs{IV, RV} where {RT<:Number, RV<:AbstractVector{RT}, IV<:(AbstractVector{<:Integer})}

neumann_bcs::NeumannBCs

times::TimeStepper{RV} where {RT<:Number, RV<:AbstractVector{RT}}

physics::NamedTuple

properties::NamedTuple

state_old::NamedTuple

state_new::NamedTuple

h1_coords::H1Field{RT, RV} where {RT<:Number, RV<:AbstractVector{RT}}

h1_field::H1Field{RT, RV} where {RT<:Number, RV<:AbstractVector{RT}}

h1_field_old::H1Field{RT, RV} where {RT<:Number, RV<:AbstractVector{RT}}

h1_hvp_scratch_field::H1Field{RT, RV} where {RT<:Number, RV<:AbstractVector{RT}}

FiniteElementContainers.Parameters — Method

Parameters(
    mesh,
    assembler,
    physics,
    properties,
    ics,
    dirichlet_bcs,
    neumann_bcs,
    times
) -> Parameters{RT, _A, IV, ICFuncs, DBCFuncs, NBCs, Phys, Props, S} where {RT<:Number, _A<:AbstractVector{RT}, IV<:(AbstractVector{<:Integer}), ICFuncs<:NamedTuple, DBCFuncs<:NamedTuple, NBCs<:(NeumannBCs{BCCaches, BCFuncs} where {BCCaches<:(NamedTuple{_A, <:Tuple{Vararg{FiniteElementContainers.NeumannBCContainer}}} where _A), BCFuncs<:(NamedTuple{_A, <:Tuple{Vararg{Function}}} where _A)}), Phys<:NamedTuple, Props<:NamedTuple, S<:(NamedTuple{_A, Tuple{Vararg{Array{Float64, 3}, N}}} where {_A, N})}

FiniteElementContainers.initialize! — Method

initialize!(p::Parameters)

FiniteElementContainers.update_bc_values! — Method

update_bc_values!(p::Parameters)

This method is used to update the stored bc values. This should be called at the beginning of any load step

TODO need to incorporate other bcs besides H1 spaces TODO need to incorporate neumann bc updates

FiniteElementContainers.update_dofs! — Method

update_dofs!(
    asm::FiniteElementContainers.AbstractAssembler,
    p::Parameters
)

FiniteElementContainers.update_time! — Method

update_time!(p::Parameters)